+++ test-parts/07-controller_files/files/con-settings.js
@@ -5,7 +5,6 @@ const chaiThings = require('chai-things');
 const commonPaths = require("../../../app/paths/files/app-paths");
 const foxPath = require(commonPaths.foxRelative);
 const commonFunctions = require(commonPaths.testCommon);
-const objectFunctions = require(commonPaths.testObject);
 const loadFoxFile = require(commonPaths.loadFox);
 const controllerSettingsFile = loadFoxFile(foxPath.settingsFile);
 
@@ -39,14 +38,12 @@ function handleFlagProperties()
 	{
 		it("Production (isProd)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'isProd');
-			objectFunctions.testPropType(controllerSettingsFile, 'isProd', 'boolean');
+			commonFunctions.testBoolean(controllerSettingsFile.isProd);
 		});
 		
 		it("Development (isDev)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'isDev');
-			objectFunctions.testPropType(controllerSettingsFile, 'isDev', 'boolean');
+			commonFunctions.testBoolean(controllerSettingsFile.isDev);
 		});
 		
 	});
@@ -59,35 +56,31 @@ function handlePathProperties()
 	{
 		it("User Storage Folder (userStoragePath)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'userStoragePath');
 			commonFunctions.testString(controllerSettingsFile.userStoragePath);
 		});
 		
 		it("Database Folder (dbsPath)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'dbsPath');
 			commonFunctions.testString(controllerSettingsFile.dbsPath);
 		});
 		
 		it("Log Folder (logsPath)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'logsPath');
 			commonFunctions.testString(controllerSettingsFile.logsPath);
 		});
 		
 		it("Log File (logFile)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'logFile');
 			commonFunctions.testString(controllerSettingsFile.logFile);
 		});
 		
 		it("Node RED Folder (flowsPath)", function()
 		{
-			objectFunctions.testPropExists(controllerSettingsFile, 'flowsPath');
 			commonFunctions.testString(controllerSettingsFile.flowsPath);
 		});
 		
 	});
 }
+
 
 module.exports = testControllerSettings;