+++ test-parts/07-controller_files/sub/rio-factory_return.js
@@ -2,22 +2,15 @@ const chai = require("chai");
 const expect = require("chai").expect;
 const commonPaths = require("../../../app/paths/files/app-paths");
 const commonFunctions = require(commonPaths.testCommon);
-const objectFunctions = require(commonPaths.testObject);
 
 
 function checkFactoryReturnValid(factReturnObject)
-{
-	objectFunctions.testPropExists(factReturnObject, 'disableRio');
-	objectFunctions.testPropExists(factReturnObject, 'getRioDeviceProperties');
-	objectFunctions.testPropExists(factReturnObject, 'registerNodeCallback');
-	objectFunctions.testPropExists(factReturnObject, 'setOutputFromNode');
-	objectFunctions.testPropExists(factReturnObject, 'getCommsErrors');
-			
-	objectFunctions.testPropType(factReturnObject, 'disableRio', 'function');
-	objectFunctions.testPropType(factReturnObject, 'getRioDeviceProperties', 'function');
-	objectFunctions.testPropType(factReturnObject, 'registerNodeCallback', 'function');
-	objectFunctions.testPropType(factReturnObject, 'setOutputFromNode', 'function');
-	objectFunctions.testPropType(factReturnObject, 'getCommsErrors', 'function');
+{	
+	commonFunctions.testFunction(factReturnObject.disableRio);
+	commonFunctions.testFunction(factReturnObject.getRioDeviceProperties);
+	commonFunctions.testFunction(factReturnObject.registerNodeCallback);
+	commonFunctions.testFunction(factReturnObject.setOutputFromNode);
+	commonFunctions.testFunction(factReturnObject.getCommsErrors);
 }